requirements.txt
accelerate==1.7.0
albumentations==2.0.7
apex==0.9.10dev
attrdict==2.0.1
decord==0.6.0
diffusers==0.33.1
einops==0.3.0
fairscale==0.4.13
faiss-cpu # pip install faiss-gpu?
fire==0.7.0
ftfy==6.3.1
gradio==6.0.1
horovod==0.28.1
huggingface_hub
imageio==2.37.0
invisible_watermark==0.2.0
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.7
keras==3.12.0
kornia==0.6.0
matplotlib==3.10.7
#mmcv.egg==info
mmdet==3.3.0
mmengine==0.10.6
mmengine_lite==0.10.7
mmsegmentation==1.2.2
natsort==8.4.0
nltk==3.9.1
numpy==2.1.3
omegaconf==2.1.1
openai==2.8.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
opendatasets==0.1.22
packaging
pandas==2.3.3
panoptic
peft==0.10.0
Pillow
ppo==0.1.1
protobuf
pudb==2019.2
pycocoevalcap==1.2
pycocotools==2.0.8
pytest==8.3.5
python_magic==0.4.27
pytorch_lightning==2.5.6
pytorch_triton_rocm
PyYAML==6.0.3
qwen_vl_utils==0.0.14
regex==2024.11.6
Requests==2.32.5
safetensors==0.7.0
scann==1.4.2
scikit_learn==1.7.2
scipy
setuptools==78.1.1
six==1.17.0
scikit-image
spacy==3.8.8
streamlit==1.45.1
tensorflow==2.19.1
tensorflow_hub==0.16.1
tensorflow_text==2.19.0
timm==1.0.22
tqdm==4.67.1
transformers==4.52.0
triton==3.3.0
vertexai==1.71.1
webdataset==1.0.2
xtuner==0.1.23
#flash_attn==2.8.3 pip install flash-attn --no-build-isolation
#torch==2.7.0
#torchvision==0.22.0